requirements.txt
bidict==0.22.0
cffi==1.15.1
click==8.0.4
cryptography==37.0.4
dnspython==2.2.1
eventlet==0.33.1
Flask==2.0.3
Flask-Cors==3.0.10
Flask-MySQLdb==1.0.1
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
mysqlclient==2.1.1
pycparser==2.21
PyJWT==2.4.0
python-engineio==3.13.2
python-socketio==4.6.0
six==1.16.0
SQLAlchemy==1.4.39
typing_extensions==4.2.0
Werkzeug==2.0.3
zipp==3.8.0
